# Conflicts:
#	.idea/artifacts/DriverApplication_war_exploded.xml
#	.idea/artifacts/GovernmentAdmin_war_exploded.xml
#	.idea/artifacts/InvoiceSystem_war_exploded.xml
#	.idea/artifacts/MovementProxy_war_exploded.xml
#	.idea/dataSources.xml
#	InvoiceSystem/InvoiceSystem/pom.xml
#	InvoiceSystem/InvoiceSystem/src/main/java/com/proftaak/invoicesystem/dao/VehicleProcessingDao.java
#	InvoiceSystem/InvoiceSystem/src/main/java/com/proftaak/invoicesystem/dao/VehicleProcessingDaoImpl.java
#	InvoiceSystem/InvoiceSystem/src/main/java/com/proftaak/invoicesystem/models/SquareRegion.java
#	azure-pipelines.pullrequest.yml

+++ UserSystem/UserSystem/src/main/java/com/proftaak/usersystem/dao/UserDao.java
@@ -1,15 +1,14 @@
 package com.proftaak.usersystem.dao;
 
 import com.proftaak.usersystem.models.ClientUser;
-
-import javax.ws.rs.client.Client;
 import com.proftaak.usersystem.models.UserVehicle;
+
 import java.util.List;
 
 public interface UserDao {
     ClientUser saveUserInformation(String name, String email, String address, String residence);
     ClientUser getClientUserByName(String name);
-    ClientUser getClientUserById(int id);
+    ClientUser getClientUserById(long id);
 	List<ClientUser> getAll();
 	ClientUser editUser(ClientUser user);
 	UserVehicle getUserVehicle(long vehicleId, long userId);
